RolePermission is now default for API endpoints
