+++ InvenTree/part/api.py
@@ -29,7 +29,7 @@ from . import serializers as part_serializers
 
 from InvenTree.views import TreeSerializer
 from InvenTree.helpers import str2bool, isNull
-from InvenTree.api import AttachmentMixin, RolePermission
+from InvenTree.api import AttachmentMixin
 
 from InvenTree.status_codes import BuildStatus
 
@@ -107,10 +107,6 @@ class CategoryList(generics.ListCreateAPIView):
     ]
 
     role_required = 'part'
-
-    permission_classes = [
-        RolePermission,
-    ]
 
 
 class CategoryDetail(generics.RetrieveUpdateDestroyAPIView):
